cnet/requirements.txt
certifi==2021.10.8
cycler==0.11.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.3
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
torch==1.10.0+rocm4.2
torchvision==0.11.1+rocm4.2
typing-extensions==3.10.0.2
